Feat : comment 유저인증 추가

+++ src/main/java/com/example/newspeed/service/CommentService.java
@@ -1,12 +1,18 @@
 package com.example.newspeed.service;
 
+import com.example.newspeed.dto.AuthUserDto;
 import com.example.newspeed.dto.comment.*;
 import com.example.newspeed.entity.Comment;
 import com.example.newspeed.entity.Post;
+import com.example.newspeed.entity.Users;
+import com.example.newspeed.exception.exceptions.AuthenticationException;
 import com.example.newspeed.exception.exceptions.NotFoundException;
 import com.example.newspeed.repository.CommentRepository;
 import com.example.newspeed.repository.PostRepository;
+import com.example.newspeed.repository.UsersRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,11 +28,13 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
+    private final UsersRepository userRepository;
     private final ModelMapper modelMapper;
 
-    public CommentService(CommentRepository commentRepository, PostRepository postRepository, ModelMapper modelMapper) {
+    public CommentService(CommentRepository commentRepository, PostRepository postRepository, UsersRepository userRepository, ModelMapper modelMapper) {
         this.commentRepository = commentRepository;
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
         this.modelMapper = modelMapper;
     }
 
@@ -35,17 +43,19 @@ public class CommentService {
      *
      * @param postId        댓글이 등록될 게시글
      * @param requestDto    요청 DTO
-     * @return {@link CommentCreateResponseDto} 반환 DTO
+     * @return {@link CommentCreateResponseDtoComment} 반환 DTO
      */
-    // TODO 회원, 게시글 기능 추가 시 회원, 게시글 정보 추가
     @Transactional
-    public CommentCreateResponseDto createComment(Long postId, CommentCreateRequestDto requestDto) {
+    public CommentCreateResponseDtoComment createComment(Long postId, CommentCreateRequestDto requestDto) {
         Post findPost = postRepository.findById(postId).orElseThrow(() -> new NotFoundException("Not Found Post"));
 
-        Comment comment = new Comment(requestDto.getContent(), findPost);
+        AuthUserDto authUserDto = getAuthUser();
+        Users user = userRepository.findById(authUserDto.getId()).orElseThrow(() -> new NotFoundException("Not Found User"));
+
+        Comment comment = new Comment(requestDto.getContent(), findPost, user);
         commentRepository.save(comment);
 
-        CommentCreateResponseDto responseDto = modelMapper.map(comment, CommentCreateResponseDto.class);
+        CommentCreateResponseDtoComment responseDto = modelMapper.map(comment, CommentCreateResponseDtoComment.class);
         return responseDto;
     }
 
@@ -53,15 +63,15 @@ public class CommentService {
      * 게시글의 모든 댓글 조회
      *
      * @param postId 조회할 게시글 ID
-     * @return {@link CommentFindResponseDto} 반환 DTO 리스트
+     * @return {@link CommentFindResponseDtoComment} 반환 DTO 리스트
      */
-    public List<CommentFindResponseDto> findCommentByPostId(Long postId) {
+    public List<CommentFindResponseDtoComment> findCommentByPostId(Long postId) {
         Post findPost = postRepository.findById(postId).orElseThrow(() -> new NotFoundException("Not Found Post"));
         List<Comment> comments = findPost.getComments();
 
-        List<CommentFindResponseDto> responseDtoList = comments
+        List<CommentFindResponseDtoComment> responseDtoList = comments
                 .stream()
-                .map(comment -> modelMapper.map(comment, CommentFindResponseDto.class))
+                .map(comment -> modelMapper.map(comment, CommentFindResponseDtoComment.class))
                 .toList();
         return responseDtoList;
     }
@@ -71,31 +81,58 @@ public class CommentService {
      *
      * @param commentId     수정할 댓글 ID
      * @param requestDto    {@link CommentUpdateRequestDto} 요청 DTO
-     * @return {@link CommentUpdateResponseDto} 반환 DTO
+     * @return {@link CommentUpdateResponseDtoComment} 반환 DTO
      */
     //TODO 회원 기능 구현 후 이름 적용 필요
-    public CommentUpdateResponseDto updateComment(Long commentId, CommentUpdateRequestDto requestDto) {
+    public CommentUpdateResponseDtoComment updateComment(Long commentId, CommentUpdateRequestDto requestDto) {
         Comment findComment = commentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("Not Found Comment"));
-        String prevContent = findComment.getContent();
+        verifyWriterAuthorities(findComment);
 
+        String prevContent = findComment.getContent();
         if(prevContent.equals(requestDto.getContent()))
             throw new IllegalArgumentException("변경할 내용이 없습니다.");
 
         findComment.updateContent(requestDto.getContent());
         commentRepository.save(findComment);
 
-        CommentUpdateResponseDto responseDto = modelMapper.map(findComment, CommentUpdateResponseDto.class);
+        CommentUpdateResponseDtoComment responseDto = modelMapper.map(findComment, CommentUpdateResponseDtoComment.class);
         responseDto.setPrevContent(prevContent);
 
         return responseDto;
     }
 
-    public CommentRemoveResponseDto deleteComment(Long commentId) {
+    public CommentRemoveResponseDtoComment deleteComment(Long commentId) {
         Comment findComment = commentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("Not Found Comment"));
+        verifyWriterAuthorities(findComment);
 
-        CommentRemoveResponseDto responseDto = modelMapper.map(findComment, CommentRemoveResponseDto.class);
+        CommentRemoveResponseDtoComment responseDto = modelMapper.map(findComment, CommentRemoveResponseDtoComment.class);
         commentRepository.delete(findComment);
 
         return responseDto;
+    }
+
+    private void verifyWriterAuthorities(Comment comment) {
+        Optional<Authentication> authentication = Optional.ofNullable(
+                SecurityContextHolder.getContext().getAuthentication()
+        );
+        if(authentication.isPresent() && authentication.get().getPrincipal() instanceof AuthUserDto) {
+            AuthUserDto auth = (AuthUserDto) authentication.get().getPrincipal();
+            if(!auth.getEmail().equals(comment.getUser().getEmail())) {
+                throw new AuthenticationException("Unauthorized");
+            }
+        } else {
+            throw new AuthenticationException("Unauthorized");
+        }
+    }
+
+    private AuthUserDto getAuthUser() {
+        Optional<Authentication> authentication = Optional.ofNullable(
+                SecurityContextHolder.getContext().getAuthentication()
+        );
+        if(authentication.isPresent() && authentication.get().getPrincipal() instanceof AuthUserDto) {
+            return (AuthUserDto) authentication.get().getPrincipal();
+        } else  {
+            throw new AuthenticationException("Unauthorized");
+        }
     }
 }
